http: $http.get does not require the full URL
The AngularJS method $http.get does not require a full path.
Simplify the Javascript code. No need to read the configuration
file in gulp and inject this value in the final scripts.

As a side effect this will permet to run the SimpleHTTPServer on
an arbitrary port without requiring the modification of the JSON
configuration file.

Signed-off-by: madrisan <davide.madrisan@gmail.com>