html viewer: first simple working multiple scan viewer
A multi scan viewer is now available.
Once the JSON scan files have all been saved in a directory,
you can run the html python viewer with the command line option '-d':

   oocs-htmlviewer.py -d <json-dir>

The first page will show a clickable list of all the available scan
reports.  This solution it's not the better one with long lists.
This point will be addressed in future commits.

Signed-off-by: madrisan <davide.madrisan@gmail.com>